PDFBOX-5660: sonar fix
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1927645 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/font/PDCIDFontType2.java
@@ -337,7 +337,7 @@ public class PDCIDFontType2 extends PDCIDFont
     {
         // todo: really we want the BBox, (for text extraction:)
         return (ttf.getHorizontalHeader().getAscender() + -ttf.getHorizontalHeader().getDescender())
-                / ttf.getUnitsPerEm(); // todo: shouldn't this be the yMax/yMin?
+                / (float) ttf.getUnitsPerEm(); // todo: shouldn't this be the yMax/yMin?
     }
 
     @Override
